+++ demo.py
@@ -1,20 +1,22 @@
-from thipster.auth.google import GoogleAuth
-import thipster.engine.engine as eng
-from thipster.parser.parser_factory import ParserFactory
-
-from thipster.parser.dsl_parser.parser import DSLParserPathNotFound
-from thipster.parser.dsl_parser.token_parser import DSLSyntaxException
-from thipster.repository.local import LocalRepo
-from thipster.terraform.cdk import CDK, CDKException
+import thipster
+from thipster.auth import Google
+from thipster.parser import ParserFactory
+from thipster.parser.dsl_parser.exceptions import (
+    DSLParserPathNotFound,
+    DSLSyntaxException,
+)
+from thipster.repository import LocalRepo
+from thipster.terraform import Terraform
+from thipster.terraform.exceptions import CDKException
 
 
 def demo():
     """Script to launch the sprint demo
     """
     file = input()
-    engine = eng.Engine(
+    engine = thipster.Engine(
         ParserFactory(), LocalRepo('/home/rcattin/THipster/tests/resources/e2e/models'),
-        GoogleAuth, CDK(),
+        Google, Terraform(),
     )
 
     try:
