+++ pyproject.toml
@@ -27,3 +27,6 @@ major_on_zero = false
 
 [pytest]
 mock_use_standalone_module = true
+
+[tool.ruff]
+exclude = ["__init__.py"]
