+++ thipster/__init__.py
@@ -1,0 +1,5 @@
+"""Thipster is a tool that enhances the compute capabilities of Terraform and \
+simplifies its file writing through its dedicated language
+"""
+
+from .engine import Engine
