+++ thipster/auth/__init__.py
@@ -3,3 +3,5 @@
 Allows users to authenticate themselves to the desired supported cloud provider.
 Currently, only Google Cloud Platform is available.
 """
+
+from .google import GoogleAuth as Google
