+++ thipster/engine/__init__.py
@@ -3,3 +3,9 @@
 This module contains the Engine class, which is the central class of the THipster
 package as well as the interfaces for the various components of the application.
 """
+
+from .engine import Engine
+from .i_auth import I_Auth
+from .i_parser import I_Parser
+from .i_repository import I_Repository
+from .i_terraform import I_Terraform
