+++ thipster/engine/i_repository.py
@@ -1,4 +1,5 @@
 from abc import ABC, abstractmethod
+
 from thipster.engine.resource_model import ResourceModel
