+++ thipster/parser/dsl_parser/__init__.py
@@ -1,0 +1,1 @@
+from .parser import DSLParser
