+++ thipster/repository/github.py
@@ -1,8 +1,9 @@
 """GithubRepo.py module
 """
 
-from thipster.repository.json import JSONRepo
 import requests
+
+from .json import JSONRepo
 
 
 class GithubRepo(JSONRepo):
