+++ thipster/terraform/__init__.py
@@ -5,3 +5,5 @@ generation of the Terraform CDK constructs and the creation of the
 Terraform code itself.
 It also allows user to execute Terraform commands from the THipster package.
 """
+
+from .cdk import CDK as Terraform
